Agregada versionado en la api. Correcciones menores en devolucion de errores.
