+++ club_cms/app/Http/Controllers/Api/v1/CourtController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers\Api\v1;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\JsonResponse;
